Preparing code for migration and ecosystem switching

+++ backend/elixir/view/resource.py
@@ -135,7 +135,9 @@ class ResourceList(APIView):
 			# issue_function(Resource.objects.get(biotoolsID=serializer.data['biotoolsID'], visibility=1), request.user)
 			es.index(index=settings.ELASTIC_SEARCH_INDEX, doc_type='_doc', body=serializer.data)
 			notify_admins(serializer.data['biotoolsID'], 'Tool CREATE', 'A tool was created.')
-			create_ecosystem_tool(serializer.data, str(request.user))
+			
+			if settings.GITHUB_ECOSYSTEM_ON:
+				create_ecosystem_tool(serializer.data, str(request.user))
 			return Response(serializer.data, status=status.HTTP_201_CREATED)
 		return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 		
@@ -351,7 +353,9 @@ class ResourceDetail(APIView):
 			count = result['hits']['total']['value']
 			if count == 1:
 				es.index(index=settings.ELASTIC_SEARCH_INDEX, doc_type='_doc', body=serializer.data, id=result['hits']['hits'][0]['_id'])
-			update_ecosystem_tool(serializer.data, str(request.user))
+			notify_admins(serializer.data['biotoolsID'], 'Tool UPDATE', 'A tool was updated.')
+			if settings.GITHUB_ECOSYSTEM_ON:
+				update_ecosystem_tool(serializer.data, str(request.user))
 			return Response(serializer.data)
 		
 		return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
@@ -379,7 +383,9 @@ class ResourceDetail(APIView):
 			count = result['hits']['total']['value']
 			if count == 1:
 				es.delete(index=settings.ELASTIC_SEARCH_INDEX, doc_type='_doc', id=result['hits']['hits'][0]['_id'])
-			delete_ecosystem_tool(biotoolsID, str(request.user))
+			notify_admins(biotoolsID, 'Tool DELETE', 'A tool was was deleted.')
+			if settings.GITHUB_ECOSYSTEM_ON:
+				delete_ecosystem_tool(biotoolsID, str(request.user))
 			return Response(status=status.HTTP_204_NO_CONTENT)
 		else:
 			return Response({"detail": "Only a superuser can remove a resource."}, status=status.HTTP_403_FORBIDDEN)
